requirements.txt
apache-beam[gcp]==2.58.0
google-cloud-bigquery==3.10.0
google-cloud-storage==2.14.0
pytest==8.3.0
